+++ src/modules/project/pages/projectCreation/views/start/LaunchStart.tsx
@@ -1,5 +1,7 @@
 import { VStack } from '@chakra-ui/react'
 
+import { ProjectCreationReferralCapture } from '../../components/ProjectCreationReferralCapture.tsx'
+
 import { FundingStyleSection } from './sections/FundingStyleSection.tsx'
 import { GetDiscoveredSection } from './sections/GetDiscoveredSection.tsx'
 import { HeroSection } from './sections/HeroSection.tsx'
@@ -13,16 +15,19 @@ import { WelcomeSection } from './sections/WelcomeSection.tsx'
 /** Launch your project landing page component */
 export const LaunchStart = () => {
   return (
-    <VStack spacing={0} paddingBottom={20} width="100%" align="center">
-      <HeroSection />
-      <WelcomeSection />
-      <SuccessStoriesSection />
-      <FundingStyleSection />
-      <GetDiscoveredSection />
-      <HowToLaunchSection />
-      <LaunchToolsSection />
-      <TipsStoriesSection />
-      <SocialFooterSection />
-    </VStack>
+    <>
+      <ProjectCreationReferralCapture />
+      <VStack spacing={0} paddingBottom={20} width="100%" align="center">
+        <HeroSection />
+        <WelcomeSection />
+        <SuccessStoriesSection />
+        <FundingStyleSection />
+        <GetDiscoveredSection />
+        <HowToLaunchSection />
+        <LaunchToolsSection />
+        <TipsStoriesSection />
+        <SocialFooterSection />
+      </VStack>
+    </>
   )
 }